+++ example/lib/widgets/week_view_widget.dart
@@ -16,7 +16,7 @@ class WeekViewWidget extends StatelessWidget {
       width: width,
       showWeekends: true,
       showLiveTimeLineInAllDays: true,
-      eventArranger: SideEventArranger(maxWidth: 30),
+      eventArranger: SideEventArranger(),
       timeLineWidth: 65,
       scrollPhysics: const BouncingScrollPhysics(),
       liveTimeIndicatorSettings: LiveTimeIndicatorSettings(