+++ lib/src/event_arrangers/event_arrangers.dart
@@ -4,14 +4,13 @@
 
 import 'dart:math' as math;
 
-import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
 
 import '../calendar_event_data.dart';
 import '../constants.dart';
 import '../extensions.dart';
 
 part 'merge_event_arranger.dart';
-
 part 'side_event_arranger.dart';
 
 /// {@template event_arranger_arrange_method_doc}